+++ R/p_direction.R
@@ -13,10 +13,16 @@
 #'   such as `"kernel"`, `"logspline"` or `"KernSmooth"`. See details.
 #' @param null The value considered as a "null" effect. Traditionally 0, but
 #'   could also be 1 in the case of ratios of change (OR, IRR, ...).
+#' @param as_p If `TRUE`, the p-direction (pd) values are converted to a
+#' frequentist p-value using [`pd_to_p()`].
+#' @param remove_na Should missing values be removed before computation? Note
+#' that `Inf` (infinity) are *not* removed.
+#' @param rvar_col Name of an `rvar`-type column. If `NULL`, each column in the
+#' data frame is assumed to represent draws from a posterior distribution.
 #' @inheritParams hdi
 #'
-#' @details
-#' ## What is the *pd*?
+#' @section What is the *pd*?:
+#'
 #' The Probability of Direction (pd) is an index of effect existence, representing
 #' the certainty with which an effect goes in a particular direction (i.e., is
 #' positive or negative / has a sign), typically ranging from 0.5 to 1 (but see
@@ -31,7 +37,8 @@
 #'   be used to draw parallels and give some reference to readers non-familiar
 #'   with Bayesian statistics (Makowski et al., 2019).
 #'
-#' ## Relationship with the p-value
+#' @section Relationship with the p-value:
+#'
 #' In most cases, it seems that the *pd* has a direct correspondence with the
 #' frequentist one-sided *p*-value through the formula (for two-sided *p*):
 #' \deqn{p = 2 \times (1 - p_d)}{p = 2 * (1 - pd)}
@@ -39,11 +46,12 @@
 #' correspond approximately to a *pd* of `95%`, `97.5%`, `99.5%` and `99.95%`.
 #' See [pd_to_p()] for details.
 #'
-#' ## Possible Range of Values
+#' @section Possible Range of Values:
+#'
 #' The largest value *pd* can take is 1 - the posterior is strictly directional.
 #' However, the smallest value *pd* can take depends on the parameter space
 #' represented by the posterior.
-#' \cr\cr
+#'
 #' **For a continuous parameter space**, exact values of 0 (or any point null
 #' value) are not possible, and so 100% of the posterior has _some_ sign, some
 #' positive, some negative. Therefore, the smallest the *pd* can be is 0.5 -
@@ -51,13 +59,13 @@
 #' 0.5 _cannot_ be used to support the null hypothesis (that the parameter does
 #' _not_ have a direction) is a similar why to how large p-values cannot be used
 #' to support the null hypothesis (see [`pd_to_p()`]; Makowski et al., 2019).
-#' \cr\cr
+#'
 #' **For a discrete parameter space or a parameter space that is a mixture
 #' between discrete and continuous spaces**, exact values of 0 (or any point
 #' null value) _are_ possible! Therefore, the smallest the *pd* can be is 0 -
 #' with 100% of the posterior mass on 0. Thus values close to 0 can be used to
 #' support the null hypothesis (see van den Bergh et al., 2021).
-#' \cr\cr
+#'
 #' Examples of posteriors representing discrete parameter space:
 #' - When a parameter can only take discrete values.
 #' - When a mixture prior/posterior is used (such as the spike-and-slab prior;
@@ -65,16 +73,17 @@
 #' - When conducting Bayesian model averaging (e.g., [weighted_posteriors()] or
 #'   `brms::posterior_average`).
 #'
-#' ## Methods of computation
+#' @section Methods of computation:
+#'
 #' The *pd* is defined as:
 #' \deqn{p_d = max({Pr(\hat{\theta} < \theta_{null}), Pr(\hat{\theta} > \theta_{null})})}{pd = max(mean(x < null), mean(x > null))}
-#' \cr\cr
+#'
 #' The most simple and direct way to compute the *pd* is to compute the
 #' proportion of positive (or larger than `null`) posterior samples, the
 #' proportion of negative (or smaller than `null`) posterior samples, and take
 #' the larger of the two. This "simple" method is the most straightforward, but
 #' its precision is directly tied to the number of posterior draws.
-#' \cr\cr
+#'
 #' The second approach relies on [density estimation][estimate_density]: It starts by
 #' estimating the continuous-smooth density function (for which many methods are
 #' available), and then computing the [area under the curve][area_under_curve]
@@ -99,7 +108,7 @@
 #'   (2021). A cautionary note on estimating effect size. Advances in Methods
 #'   and Practices in Psychological Science, 4(1). \doi{10.1177/2515245921992035}
 #'
-#' @examples
+#' @examplesIf requireNamespace("rstanarm", quietly = TRUE) && requireNamespace("emmeans", quietly = TRUE) && requireNamespace("brms", quietly = TRUE) && requireNamespace("BayesFactor", quietly = TRUE)
 #' library(bayestestR)
 #'
 #' # Simulate a posterior distribution of mean 1 and SD 1
@@ -116,36 +125,28 @@
 #' \donttest{
 #' # rstanarm models
 #' # -----------------------------------------------
-#' if (require("rstanarm")) {
-#'   model <- rstanarm::stan_glm(mpg ~ wt + cyl,
-#'     data = mtcars,
-#'     chains = 2, refresh = 0
-#'   )
-#'   p_direction(model)
-#'   p_direction(model, method = "kernel")
-#' }
+#' model <- rstanarm::stan_glm(mpg ~ wt + cyl,
+#'   data = mtcars,
+#'   chains = 2, refresh = 0
+#' )
+#' p_direction(model)
+#' p_direction(model, method = "kernel")
 #'
 #' # emmeans
 #' # -----------------------------------------------
-#' if (require("emmeans")) {
-#'   p_direction(emtrends(model, ~1, "wt", data = mtcars))
-#' }
+#' p_direction(emmeans::emtrends(model, ~1, "wt", data = mtcars))
 #'
 #' # brms models
 #' # -----------------------------------------------
-#' if (require("brms")) {
-#'   model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
-#'   p_direction(model)
-#'   p_direction(model, method = "kernel")
-#' }
+#' model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
+#' p_direction(model)
+#' p_direction(model, method = "kernel")
 #'
 #' # BayesFactor objects
 #' # -----------------------------------------------
-#' if (require("BayesFactor")) {
-#'   bf <- ttestBF(x = rnorm(100, 1, 1))
-#'   p_direction(bf)
-#'   p_direction(bf, method = "kernel")
-#' }
+#' bf <- BayesFactor::ttestBF(x = rnorm(100, 1, 1))
+#' p_direction(bf)
+#' p_direction(bf, method = "kernel")
 #' }
 #' @export
 p_direction <- function(x, ...) {
@@ -165,9 +166,21 @@ p_direction.default <- function(x, ...) {
 
 #' @rdname p_direction
 #' @export
-p_direction.numeric <- function(x, method = "direct", null = 0, ...) {
+p_direction.numeric <- function(x,
+                                method = "direct",
+                                null = 0,
+                                as_p = FALSE,
+                                remove_na = TRUE,
+                                ...) {
   obj_name <- insight::safe_deparse_symbol(substitute(x))
-  out <- p_direction(data.frame(Posterior = x), method = method, null = null, ...)
+  out <- p_direction(
+    data.frame(Posterior = x),
+    method = method,
+    null = null,
+    as_p = as_p,
+    remove_na = remove_na,
+    ...
+  )
   attr(out, "object_name") <- obj_name
   out
 }
@@ -175,14 +188,78 @@ p_direction.numeric <- function(x, method = "direct", null = 0, ...) {
 
 #' @rdname p_direction
 #' @export
-p_direction.data.frame <- function(x, method = "direct", null = 0, ...) {
+p_direction.data.frame <- function(x,
+                                   method = "direct",
+                                   null = 0,
+                                   as_p = FALSE,
+                                   remove_na = TRUE,
+                                   rvar_col = NULL,
+                                   ...) {
   obj_name <- insight::safe_deparse_symbol(substitute(x))
+
+  if (is.null(rvar_col)) {
+    return(.p_direction_df(
+      x,
+      method = method,
+      null = null,
+      as_p = as_p,
+      remove_na = remove_na,
+      obj_name = obj_name,
+      ...
+    ))
+  }
+
+  if (length(rvar_col) != 1L && !rvar_col %in% colnames(x)) {
+    insight::format_error("The `rvar_col` argument must be a single, valid column name.")
+  }
+
+  out <- p_direction(
+    x[[rvar_col]],
+    method = method,
+    null = null,
+    as_p = as_p,
+    remove_na = remove_na,
+    ...
+  )
+
+  x[["pd"]] <- out[["pd"]]
+  attr(x, "object_name") <- obj_name
+  attr(x, "as_p") <- as_p
+
+  x
+}
+
+
+#' @keywords internal
+.p_direction_df <- function(x,
+                            method = "direct",
+                            null = 0,
+                            as_p = FALSE,
+                            remove_na = TRUE,
+                            obj_name = NULL,
+                            ...) {
   x <- .select_nums(x)
 
   if (ncol(x) == 1) {
-    pd <- .p_direction(x[[1]], method = method, null = null, ...)
+    pd <- .p_direction(
+      x[[1]],
+      method = method,
+      null = null,
+      as_p = as_p,
+      remove_na = remove_na,
+      ...
+    )
   } else {
-    pd <- sapply(x, .p_direction, method = method, null = null, simplify = TRUE, ...)
+    pd <- sapply(
+      x,
+      .p_direction,
+      method = method,
+      null = null,
+      as_p = as_p,
+      remove_na = remove_na,
+      simplify = TRUE,
+      ...
+    )
   }
 
   out <- data.frame(
@@ -192,7 +269,13 @@ p_direction.data.frame <- function(x, method = "direct", null = 0, ...) {
     stringsAsFactors = FALSE
   )
 
+  # rename column
+  if (as_p) {
+    colnames(out)[2] <- "p"
+  }
+
   attr(out, "object_name") <- obj_name
+  attr(out, "as_p") <- as_p
   class(out) <- unique(c("p_direction", "see_p_direction", class(out)))
 
   out
@@ -200,8 +283,20 @@ p_direction.data.frame <- function(x, method = "direct", null = 0, ...) {
 
 
 #' @export
-p_direction.draws <- function(x, method = "direct", null = 0, ...) {
-  p_direction(.posterior_draws_to_df(x), method = method, null = null, ...)
+p_direction.draws <- function(x,
+                              method = "direct",
+                              null = 0,
+                              as_p = FALSE,
+                              remove_na = TRUE,
+                              ...) {
+  p_direction(
+    .posterior_draws_to_df(x),
+    method = method,
+    null = null,
+    as_p = as_p,
+    remove_na = remove_na,
+    ...
+  )
 }
 
 #' @export
@@ -210,11 +305,18 @@ p_direction.rvar <- p_direction.draws
 
 #' @rdname p_direction
 #' @export
-p_direction.MCMCglmm <- function(x, method = "direct", null = 0, ...) {
+p_direction.MCMCglmm <- function(x,
+                                 method = "direct",
+                                 null = 0,
+                                 as_p = FALSE,
+                                 remove_na = TRUE,
+                                 ...) {
   nF <- x$Fixed$nfl
   out <- p_direction(as.data.frame(x$Sol[, 1:nF, drop = FALSE]),
     method = method,
     null = null,
+    as_p = as_p,
+    remove_na = remove_na,
     ...
   )
   attr(out, "object_name") <- insight::safe_deparse_symbol(substitute(x))
@@ -223,20 +325,32 @@ p_direction.MCMCglmm <- function(x, method = "direct", null = 0, ...) {
 
 
 #' @export
-p_direction.mcmc <- function(x, method = "direct", null = 0, ...) {
-  p_direction(as.data.frame(x), method = method, null = null, ...)
+p_direction.mcmc <- function(x,
+                             method = "direct",
+                             null = 0,
+                             as_p = FALSE,
+                             remove_na = TRUE,
+                             ...) {
+  p_direction(
+    as.data.frame(x),
+    method = method,
+    null = null,
+    as_p = as_p,
+    remove_na = remove_na,
+    ...
+  )
 }
 
 
 #' @export
-p_direction.BGGM <- function(x, method = "direct", null = 0, ...) {
-  p_direction(as.data.frame(x), method = method, null = null, ...)
+p_direction.BGGM <- function(x, method = "direct", null = 0, as_p = FALSE, remove_na = TRUE, ...) {
+  p_direction(as.data.frame(x), method = method, null = null, as_p = as_p, remove_na = remove_na, ...)
 }
 
 
 #' @export
-p_direction.bcplm <- function(x, method = "direct", null = 0, ...) {
-  p_direction(insight::get_parameters(x), method = method, null = null, ...)
+p_direction.bcplm <- function(x, method = "direct", null = 0, as_p = FALSE, remove_na = TRUE, ...) {
+  p_direction(insight::get_parameters(x), method = method, null = null, as_p = as_p, remove_na = remove_na, ...)
 }
 
 #' @export
@@ -253,6 +367,8 @@ p_direction.bayesQR <- p_direction.bcplm
 p_direction.bamlss <- function(x,
                                method = "direct",
                                null = 0,
+                               as_p = FALSE,
+                               remove_na = TRUE,
                                component = c("all", "conditional", "location"),
                                ...) {
   component <- match.arg(component)
@@ -260,6 +376,8 @@ p_direction.bamlss <- function(x,
     insight::get_parameters(x, component = component),
     method = method,
     null = null,
+    as_p = as_p,
+    remove_na = remove_na,
     ...
   )
   out <- .add_clean_parameters_attribute(out, x)
@@ -269,10 +387,10 @@ p_direction.bamlss <- function(x,
 
 #' @rdname p_direction
 #' @export
-p_direction.emmGrid <- function(x, method = "direct", null = 0, ...) {
+p_direction.emmGrid <- function(x, method = "direct", null = 0, as_p = FALSE, remove_na = TRUE, ...) {
   xdf <- insight::get_parameters(x)
 
-  out <- p_direction(xdf, method = method, null = null, ...)
+  out <- p_direction(xdf, method = method, null = null, as_p = as_p, remove_na = remove_na, ...)
   attr(out, "object_name") <- insight::safe_deparse_symbol(substitute(x))
   out
 }
@@ -289,6 +407,8 @@ p_direction.emm_list <- p_direction.emmGrid
                                 parameters,
                                 method = "direct",
                                 null = 0,
+                                as_p = FALSE,
+                                remove_na = TRUE,
                                 ...) {
   p_direction(
     insight::get_parameters(
@@ -299,6 +419,8 @@ p_direction.emm_list <- p_direction.emmGrid
     ),
     method = method,
     null = null,
+    as_p = as_p,
+    remove_na = remove_na,
     ...
   )
 }
@@ -310,6 +432,8 @@ p_direction.sim.merMod <- function(x,
                                    parameters = NULL,
                                    method = "direct",
                                    null = 0,
+                                   as_p = FALSE,
+                                   remove_na = TRUE,
                                    ...) {
   effects <- match.arg(effects)
 
@@ -320,6 +444,8 @@ p_direction.sim.merMod <- function(x,
     parameters = parameters,
     method = method,
     null = null,
+    as_p = as_p,
+    remove_na = remove_na,
     ...
   )
   attr(out, "data") <- insight::get_parameters(x, effects = effects, parameters = parameters)
@@ -332,6 +458,8 @@ p_direction.sim <- function(x,
                             parameters = NULL,
                             method = "direct",
                             null = 0,
+                            as_p = FALSE,
+                            remove_na = TRUE,
                             ...) {
   out <- .p_direction_models(
     x = x,
@@ -340,6 +468,8 @@ p_direction.sim <- function(x,
     parameters = parameters,
     method = method,
     null = null,
+    as_p = as_p,
+    remove_na = remove_na,
     ...
   )
   attr(out, "data") <- insight::get_parameters(x, parameters = parameters)
@@ -356,6 +486,8 @@ p_direction.stanreg <- function(x,
                                 parameters = NULL,
                                 method = "direct",
                                 null = 0,
+                                as_p = FALSE,
+                                remove_na = TRUE,
                                 ...) {
   effects <- match.arg(effects)
   component <- match.arg(component)
@@ -371,6 +503,8 @@ p_direction.stanreg <- function(x,
       ),
       method = method,
       null = null,
+      as_p = as_p,
+      remove_na = remove_na,
       ...
     ),
     cleaned_parameters,
@@ -380,6 +514,7 @@ p_direction.stanreg <- function(x,
   attr(out, "clean_parameters") <- cleaned_parameters
   class(out) <- unique(c("p_direction", "see_p_direction", class(out)))
   attr(out, "object_name") <- insight::safe_deparse_symbol(substitute(x))
+  attr(out, "as_p") <- as_p
   out
 }
 
@@ -398,6 +533,8 @@ p_direction.brmsfit <- function(x,
                                 parameters = NULL,
                                 method = "direct",
                                 null = 0,
+                                as_p = FALSE,
+                                remove_na = TRUE,
                                 ...) {
   effects <- match.arg(effects)
   component <- match.arg(component)
@@ -413,6 +550,8 @@ p_direction.brmsfit <- function(x,
       ),
       method = method,
       null = null,
+      as_p = as_p,
+      remove_na = remove_na,
       ...
     ),
     cleaned_parameters
@@ -421,14 +560,27 @@ p_direction.brmsfit <- function(x,
   attr(out, "clean_parameters") <- cleaned_parameters
   class(out) <- unique(c("p_direction", "see_p_direction", class(out)))
   attr(out, "object_name") <- insight::safe_deparse_symbol(substitute(x))
+  attr(out, "as_p") <- as_p
   out
 }
 
 
 #' @rdname p_direction
 #' @export
-p_direction.BFBayesFactor <- function(x, method = "direct", null = 0, ...) {
-  out <- p_direction(insight::get_parameters(x), method = method, null = null, ...)
+p_direction.BFBayesFactor <- function(x,
+                                      method = "direct",
+                                      null = 0,
+                                      as_p = FALSE,
+                                      remove_na = TRUE,
+                                      ...) {
+  out <- p_direction(
+    insight::get_parameters(x),
+    method = method,
+    null = null,
+    as_p = as_p,
+    remove_na = remove_na,
+    ...
+  )
   attr(out, "object_name") <- insight::safe_deparse_symbol(substitute(x))
   out
 }
@@ -438,6 +590,8 @@ p_direction.BFBayesFactor <- function(x, method = "direct", null = 0, ...) {
 p_direction.get_predicted <- function(x,
                                       method = "direct",
                                       null = 0,
+                                      as_p = FALSE,
+                                      remove_na = TRUE,
                                       use_iterations = FALSE,
                                       verbose = TRUE,
                                       ...) {
@@ -447,6 +601,8 @@ p_direction.get_predicted <- function(x,
         as.data.frame(t(attributes(x)$iterations)),
         method = method,
         null = null,
+        as_p = as_p,
+        remove_na = remove_na,
         verbose = verbose,
         ...
       )
@@ -458,6 +614,8 @@ p_direction.get_predicted <- function(x,
     out <- p_direction(as.numeric(x),
       method = method,
       null = null,
+      as_p = as_p,
+      remove_na = remove_na,
       verbose = verbose,
       ...
     )
@@ -490,7 +648,30 @@ p_direction.parameters_model <- function(x, ...) {
 
 
 #' @keywords internal
-.p_direction <- function(x, method = "direct", null = 0, ...) {
+.p_direction <- function(x,
+                         method = "direct",
+                         null = 0,
+                         as_p = FALSE,
+                         remove_na = TRUE,
+                         ...) {
+  # handle missing values
+  if (remove_na) {
+    x <- x[!is.na(x)]
+  }
+  # sanity check
+  if (length(x) == 0) {
+    insight::format_error("No valid values found. Maybe the data contains only missing values.")
+  }
+  # sanity check
+  if (anyNA(x)) {
+    return(NA_real_)
+  }
+
+  # any inf values? then warn...
+  if (any(is.infinite(x))) {
+    insight::format_warning("Infinite values detected. These are not removed. Please check your results carefully!")
+  }
+
   if (method == "direct") {
     pdir <- max(
       length(x[x > null]), # pd positive
@@ -510,6 +691,11 @@ p_direction.parameters_model <- function(x, ...) {
     }
   }
 
+  # convert to frequentist p?
+  if (as_p) {
+    pdir <- pd_to_p(pdir)
+  }
+
   pdir
 }
 
@@ -523,7 +709,12 @@ p_direction.parameters_model <- function(x, ...) {
 #' @export
 as.numeric.p_direction <- function(x, ...) {
   if (inherits(x, "data.frame")) {
-    as.numeric(as.vector(x$pd))
+    # check if we have frequentist p-values
+    if (isTRUE(attributes(x)$as_p) && "p" %in% colnames(x)) {
+      as.numeric(as.vector(x$p))
+    } else {
+      as.numeric(as.vector(x$pd))
+    }
   } else {
     as.vector(x)
   }
@@ -533,3 +724,7 @@ as.numeric.p_direction <- function(x, ...) {
 #' @method as.double p_direction
 #' @export
 as.double.p_direction <- as.numeric.p_direction
+
+#' @method as.vector p_direction
+#' @export
+as.vector.p_direction <- as.numeric.p_direction